+++ deepctr_torch/__init__.py
@@ -2,5 +2,5 @@ from . import layers
 from . import models
 from .utils import check_version
 
-__version__ = '0.2.8'
+__version__ = '0.2.9'
 check_version(__version__)